DeviceIOCTL: fixed interface type for GET_VOLUME_SIZE from off_t to unsigned long long

+++ miosix/arch/drivers/sdmmc/lpc2000_sd.cpp
@@ -533,7 +533,7 @@ int SPISDDriver::ioctl(int cmd, void* arg)
     }
     case IOCTL_GET_VOLUME_SIZE: {
         DBG("IOCTL_GET_VOLUME_SIZE\n");
-        off_t* sizePtr=static_cast<off_t*>(arg);
+        unsigned long long* sizePtr=static_cast<unsigned long long*>(arg);
         if (sizePtr==nullptr) {
             return -EINVAL;
         }
